requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
requests
beautifulsoup4
jinja2
python-dotenv